refactor(code): simplify extension hosting
