+++ libs/code/tests/unit_tests/test_server_manager.py
@@ -549,7 +549,6 @@ class TestStartServerAndGetAgent:
         config = json.loads((tmp_path / "langgraph.json").read_text())
         assert config["graphs"]["agent"] == "./server_graph.py:make_graph"
         assert config["checkpointer"]["path"] == "./checkpointer.py:create_checkpointer"
-        assert "http" not in config
 
     def test_builtin_server_registers_only_the_agent_graph(
         self, tmp_path: Path
